refactor(crypto): derive_private_key accepts u128, drop negative branch
Production search callers only ever pass non-negative nonces; the
i128 signature and the negative-arm branch only exist to satisfy
two test cases (test_signed_nonce, test_derive_private_key_i128_min).

Switch the signature to u128, drop the branch, drop the two unit
tests, and adjust call sites in parallel.rs / bsgs.rs / kangaroo.rs /
search/mod.rs / cli.rs / benches / fixtures to use .unsigned_abs()
at the boundary. Also delete test_kangaroo_negative_start and
test_bsgs_negative_start, which depended on the removed path.

Cargo.lock is touched by full build but contains only metadata;
revert it from this commit and let the next chore commit pick it up.

+++ benches/search.rs
@@ -48,7 +48,7 @@ fn bench_derive_private_key(c: &mut Criterion) {
     let beta = Scalar::from(7u64);
 
     c.bench_function("derive_private_key", |b| {
-        b.iter(|| derive_private_key(black_box(42i128), black_box(alpha), black_box(beta)));
+        b.iter(|| derive_private_key(black_box(42u128), black_box(alpha), black_box(beta)));
     });
 }
 
@@ -69,7 +69,7 @@ fn bench_search_chunk(c: &mut Criterion) {
             let step_scalar = alpha * Scalar::from(1u64);
             let step_point = ProjectivePoint::GENERATOR * step_scalar;
 
-            let mut cur_scalar = derive_private_key(0, alpha, beta);
+            let mut cur_scalar = derive_private_key(0u128, alpha, beta);
             let mut point = ProjectivePoint::GENERATOR * cur_scalar;
 
             b.iter(|| {
@@ -91,7 +91,7 @@ fn bench_openmap(c: &mut Criterion) {
         group.bench_with_input(BenchmarkId::new("insert_get", size), size, |b, &size| {
             let mut map = OpenMap::with_capacity(size);
             for i in 0..size {
-                let key = u128::from(i);
+                let key = u128::from(u64::from(i as u32));
                 map.insert(key, i as u64);
             }
             b.iter(|| {
